# src/server/swagger/swagger.json
{
  "openapi" : "3.0.2",
  "info" : {
    "title" : "Network Detection and Response (NDR)",
    "description" : "Advanced Cybersecurity Analytics Service (ACAS) for AI-based analysis with explainable intelligence of network events captured by the Montimage Monitoring Tools (MMT)",
    "termsOfService" : "https://github.com/Montimage/acas",
    "contact" : {
      "email" : "developer@montimage.com"
    },
    "license" : {
      "name" : "Montimage",
      "url" : "https://www.montimage.eu"
    },
    "version" : "v1"
  },
...
  "servers": [
    {
      "description": "Localhost Server",
      "url": "http://localhost:31057/api/"
    },
    {
      "description": "Production Server",
      "url": "https://ndr.montimage.eu/api/"
    }
  ],
...
  "paths" : {
...
    "/features/extract" : {
      "post" : {
        "tags" : [ "features" ],
        "summary" : "Extract features from PCAP",
        "description" : "Extract network features from PCAP file",
        "operationId" : "extractFeatures",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "pcapFile" : {"type" : "string"},
                  "isMalicious" : {"type" : "boolean"},
                  "useQueue" : {"type" : "boolean", "example" : false}
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Features extracted",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "ok" : {"type" : "boolean"},
                    "sessionId" : {"type" : "string"},
                    "csvFile" : {"type" : "string"},
                    "jobId" : {"type" : "string"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/features/status/{jobId}" : {
      "get" : {
        "tags" : [ "features" ],
        "summary" : "Get feature extraction job status",
        "description" : "Get status of queued feature extraction job",
        "operationId" : "getFeatureJobStatus",
        "parameters" : [ {
          "name" : "jobId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Job status retrieved"
          }
        }
      }
    },
...
  },
...
}